Add give dialog to view online option on book details
