+++ src/app/pages/details/common/get-this-title-files/give-before-pdf/use-give-dialog.tsx
@@ -3,6 +3,7 @@ import {useDialog} from '~/components/dialog/dialog';
 import GiveBeforeOnline from './give-before-online';
 import GiveBeforePdf from './give-before-pdf';
 import useDonationPopupData from './use-donation-popup-data';
+import {isMobileDisplay} from '~/helpers/device';
 
 export default function useGiveDialog() {
     const [Dialog, open, close] = useDialog();
@@ -36,6 +37,21 @@ export default function useGiveDialog() {
         open,
         enabled: !data.hide_donation_popup
     };
+}
+
+export function useOpenGiveDialog() {
+    const {GiveDialog, open, enabled} = useGiveDialog();
+    const openGiveDialog = React.useCallback(
+        (event: React.MouseEvent) => {
+            if (enabled && !isMobileDisplay()) {
+                event.preventDefault();
+                open();
+            }
+        },
+        [enabled, open]
+    );
+
+    return {GiveDialog, openGiveDialog};
 }
 
 export type VariantOptions = 'online' | undefined;
